requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.6
django-grappelli==4.0.2
django-import-export==4.3.7
django-jazzmin==3.0.1
djangorestframework==3.15.2
dotenv==0.9.9
executing==2.1.0
filelock==3.17.0
Flask==3.1.0
flask-core==2.9.0
flask-cors==6.0.1
fonttools==4.55.3
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mysql-connector==2.2.9
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==308
pyzmq==26.2.0
qrcode==8.0
requests==2.32.3
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
